# Conflicts:
#	docs/quickstart.md
#	flinkx-core/src/main/java/com/dtstack/flinkx/util/RangeSplitUtil.java
#	flinkx-kafka/flinkx-kafka-writer/src/main/java/com/dtstack/flinkx/kafka/writer/KafkaWriter.java
#	flinkx-kafka10/flinkx-kafka10-writer/src/main/java/com/dtstack/flinkx/kafka10/writer/Kafka10Writer.java
#	flinkx-kafka11/flinkx-kafka11-writer/src/main/java/com/dtstack/flinkx/kafka11/writer/Kafka11Writer.java
#	flinkx-launcher/src/main/java/com/dtstack/flinkx/launcher/perJob/PerJobClusterClientBuilder.java
#	flinkx-launcher/src/main/resources/log4j.properties
#	flinkx-postgresql/flinkx-postgresql-core/src/main/java/com/dtstack/flinkx/postgresql/PostgresqlTypeConverter.java
#	flinkx-rdb/flinkx-rdb-core/src/main/java/com/dtstack/flinkx/rdb/util/DbUtil.java
#	flinkx-rdb/flinkx-rdb-writer/src/main/java/com/dtstack/flinkx/rdb/outputformat/JdbcOutputFormat.java
#	flinkx-test/src/main/java/com/dtstack/flinkx/test/LocalTest.java
#	pom.xml

+++ flinkx-postgresql/flinkx-postgresql-core/src/main/java/com/dtstack/flinkx/postgresql/PostgresqlTypeConverter.java
@@ -77,7 +77,7 @@ public class PostgresqlTypeConverter implements TypeConverterInterface {
             if(dataValue.contains(".")){
                 dataValue =  new BigDecimal(dataValue).stripTrailingZeros().toPlainString();
             }
-            data = Integer.parseInt(dataValue);
+            data = Long.parseLong(dataValue);
         }
 
         return data;
